Generalize status check script, use smaller test data set for now
